test: move to manual
