+++ .vscode/launch.json
@@ -8,7 +8,7 @@
             "name": "test",
             "type": "node",
             "request": "launch",
-            "program": "${workspaceRoot}/test/single.js",
+            "program": "${workspaceRoot}/test/manual/single.js",
             "stopOnEntry": false,
             "args": [],
             "cwd": "${workspaceRoot}",
@@ -33,7 +33,7 @@
             "name": "test snapshot",
             "type": "node",
             "request": "launch",
-            "program": "${workspaceRoot}/test/single.js",
+            "program": "${workspaceRoot}/test/manual/single.js",
             "stopOnEntry": false,
             "args": [],
             "cwd": "${workspaceRoot}",
